# Harvest the categorical power in `CellCollection.save_sonata`