--- a/voxcell/cell_collection.py
+++ b/voxcell/cell_collection.py
@@ -375,7 +375,11 @@ class CellCollection:
                 if name.startswith(self.SONATA_DYNAMIC_PROPERTY):
                     name = name.split(self.SONATA_DYNAMIC_PROPERTY)[1]
                     dt = str_dt if series.dtype == object else series.dtype
-                    group.create_dataset(f'dynamics_params/{name}', data=series.to_numpy(), dtype=dt)
+                    group.create_dataset(
+                        f'dynamics_params/{name}',
+                        data=series.to_numpy(),
+                        dtype=dt,
+                    )
                 elif _is_string_enum(series) or (series.dtype == object and name in forced_library):
                     indices, unique_values = series.factorize()
                     if name in forced_library or len(unique_values) < .5 * len(indices):